fix: set up qtmsg logger earlier
qt log messages were printed to console but not written to the log file
